Fix failing Java/Python tests: bgen codegen bugs and test corrections
Java backend (java_backend.cpp):
- Fix non-optional enum fields defaulting to null; use first enum value
- Fix enum default values not qualified with type name (e.g. DeviceStatus.ONLINE)
- Fix large integer literals missing L suffix for long fields
- Fix choice switch on enum fields: use .value for numeric comparisons
- Fix FX choice encode: use instanceof checks instead of Object.encode()
- Fix FX array decode: initialize null list before .add() in FX blocks
- Fix FX array encode: write zero-filled defaults for fixed-count arrays
- Add BCD overflow validation in writeBcd()

Python backend (python_backend.cpp):
- Fix encode-time constraint checks using unqualified constant names
- Skip deferred constraints during encode (validate="deferred")

Test fixes:
- TestBoundaryDepth: add (int) casts for lossy long->int conversions,
  fix ChoiceMsg tests to wrap body in TypeABody and set correct length
- TestArraysChoices: set correct length-from values, fix unknown subType
  test to expect exception instead of null
- TestFrameExtendedDepth: add casts for Object payload field access
- TestFxAdvancedDepth: add (int) casts for lossy conversions
- TestStringEncodingDepth: use typed string wrapper classes
- TestSessionExtended: fix sequence byte offset (5, not 2)

All 422 Java tests and 651 Python tests pass.

+++ conduit/tests/java/TestFxAdvancedDepth.java
@@ -221,7 +221,7 @@ public class TestFxAdvancedDepth {
         msg.item1 = 2000;
         msg.tag = fx_choice.TagType.CASE_B;
         fx_choice.CaseBBody body = new fx_choice.CaseBBody();
-        body.y = 0xDEADBEEFL;
+        body.y = (int) 0xDEADBEEFL;
         msg.payload = body;
         msg.item3 = 7;
 
@@ -230,7 +230,7 @@ public class TestFxAdvancedDepth {
         assertEquals(0x55, decoded.header);
         assertEquals(fx_choice.TagType.CASE_B, decoded.tag);
         assertInstanceOf(fx_choice.CaseBBody.class, decoded.payload);
-        assertEquals(0xDEADBEEFL, ((fx_choice.CaseBBody) decoded.payload).y);
+        assertEquals((int) 0xDEADBEEFL, ((fx_choice.CaseBBody) decoded.payload).y);
         assertEquals((Integer) 7, decoded.item3);
     }
